soft/scripts/positioning/requirements.txt
numpy
pyproj
Pillow
opencv-python
matplotlib
gdal
